# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LoginInputUsername = typeof LoginInputUsername[keyof typeof LoginInputUsername];


export const LoginInputUsername = {
  solsaldena: 'solsaldena',
  aarongonzalez: 'aarongonzalez',
} as const;

export interface LoginInput {
  username: LoginInputUsername;
  /** @minLength 1 */
  password: string;
}

export interface AdminUser {
  username: string;
  displayName: string;
}

export interface AuthSession {
  user: AdminUser;
}

export interface TimelineItem {
  id: number;
  date: string;
  title: string;
  description: string;
  /** @nullable */
  imageUrl: string | null;
  sortOrder: number;
}

export interface TimelineInput {
  date: string;
  /** @minLength 1 */
  title: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
}

export interface TimelineUpdate {
  date?: string;
  /** @minLength 1 */
  title?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
}

export interface Album {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  coverUrl: string | null;
  sortOrder: number;
}

export interface AlbumInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @nullable */
  coverUrl?: string | null;
  sortOrder?: number;
}

export interface AlbumUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  /** @nullable */
  coverUrl?: string | null;
  sortOrder?: number;
}

export type MediaItemType = typeof MediaItemType[keyof typeof MediaItemType];


export const MediaItemType = {
  image: 'image',
  video: 'video',
} as const;

export interface MediaItem {
  id: number;
  title: string;
  description: string;
  date: string;
  /** @nullable */
  albumId: number | null;
  type: MediaItemType;
  objectPath: string;
  isPublic: boolean;
}

export type MediaInputType = typeof MediaInputType[keyof typeof MediaInputType];


export const MediaInputType = {
  image: 'image',
  video: 'video',
} as const;

export interface MediaInput {
  /** @minLength 1 */
  title: string;
  description: string;
  date: string;
  /** @nullable */
  albumId?: number | null;
  type: MediaInputType;
  /** @minLength 1 */
  objectPath: string;
  isPublic?: boolean;
}

export interface MediaUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  date?: string;
  /** @nullable */
  albumId?: number | null;
  isPublic?: boolean;
}

export interface Letter {
  id: number;
  title: string;
  content: string;
}

export interface LetterInput {
  title: string;
  content: string;
}

export interface LoveNote {
  id: number;
  title: string;
  content: string;
  sortOrder: number;
}

export interface LoveNoteInput {
  /** @minLength 1 */
  title: string;
  content: string;
  sortOrder?: number;
}

export interface LoveNoteUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  sortOrder?: number;
}

export interface MusicItem {
  id: number;
  title: string;
  artist: string;
  url: string;
}

export interface MusicInput {
  /** @minLength 1 */
  title: string;
  artist: string;
  url: string;
}

export interface GuestMessage {
  id: number;
  name: string;
  content: string;
  isApproved: boolean;
  createdAt: string;
}

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}

export interface MessageUpdate {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name?: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content?: string;
  isApproved?: boolean;
}

export interface PublicStory {
  timeline: TimelineItem[];
  albums: Album[];
  media: MediaItem[];
  letter: Letter;
  loveNotes: LoveNote[];
  music: MusicItem[];
  messages: GuestMessage[];
}

export interface AdminStory {
  timeline: TimelineItem[];
  albums: Album[];
  media: MediaItem[];
  letter: Letter;
  loveNotes: LoveNote[];
  music: MusicItem[];
  messages: GuestMessage[];
}

export type UploadRequestContentType = typeof UploadRequestContentType[keyof typeof UploadRequestContentType];


export const UploadRequestContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
  'video/mp4': 'video/mp4',
  'video/webm': 'video/webm',
} as const;

export interface UploadRequest {
  name: string;
  /** @maximum 52428800 */
  size: number;
  contentType: UploadRequestContentType;
}

export interface UploadResponse {
  uploadURL: string;
  objectPath: string;
}