# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a private object storage upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadRequest"
      responses:
        "200":
          description: Presigned upload details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "401":
          description: Not authenticated
components:
...
  schemas:
...
    UploadRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string }
        size: { type: number, maximum: 52428800 }
        contentType: { type: string, enum: [image/jpeg, image/png, image/webp, video/mp4, video/webm] }
    UploadResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }