# lib/api-zod/src/generated/types/uploadRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UploadRequestContentType } from './uploadRequestContentType';

export interface UploadRequest {
  name: string;
  /** @maximum 52428800 */
  size: number;
  contentType: UploadRequestContentType;
}